requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.6.0
    # via django
async-timeout==4.0.2
    # via redis
certifi==2022.12.7
    # via requests
charset-normalizer==3.0.1
    # via requests
coreapi==2.3.3
    # via drf-yasg
coreschema==0.0.4
    # via
    #   coreapi
    #   drf-yasg
django==4.1.6
    # via
    #   -r requirements.in
    #   django-filter
    #   django-phonenumber-field
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-yasg
django-cleanup==7.0.0
    # via -r requirements.in
django-filter==22.1
    # via -r requirements.in
django-phonenumber-field[phonenumbers]==7.0.2
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-yasg
djangorestframework-simplejwt==5.2.2
    # via -r requirements.in
drf-yasg==1.21.5
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
idna==3.4
    # via requests
inflection==0.5.1
    # via drf-yasg
itypes==1.2.0
    # via coreapi
jinja2==3.1.2
    # via coreschema
markupsafe==2.1.2
    # via jinja2
packaging==23.0
    # via drf-yasg
phonenumbers==8.13.6
    # via django-phonenumber-field
psycopg2-binary==2.9.5
    # via -r requirements.in
pyjwt==2.6.0
    # via djangorestframework-simplejwt
python-decouple==3.7
    # via -r requirements.in
pytz==2022.7.1
    # via
    #   djangorestframework
    #   drf-yasg
redis==4.5.1
    # via -r requirements.in
requests==2.28.2
    # via coreapi
ruamel-yaml==0.17.21
    # via drf-yasg
ruamel-yaml-clib==0.2.7
    # via ruamel-yaml
sqlparse==0.4.3
    # via django
uritemplate==4.1.1
    # via
    #   coreapi
    #   drf-yasg
urllib3==1.26.14
    # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
